Circular spring plotter responds to stretch amount now.
Added an input to method plotCircSpring of class RunnerPlotter to modulate the appearance of the spring based on how much it is stretched.
Also added a bug fix if the user inputs a desired radius for the circular spring that is too small, it will default to the smallest circle possible.