Param study for Swing model on initial yank impulse
Starting from a coefficient of restitution of 2, which imparts a velocity to the swing foot equal to twice the final velocity of the SLIP stance leg of the previous step, and incrementing it downwards until the model can no longer run without violating constraints.  The optimizer is allowed to minimize the spring constants and the hip spring set point, but the swing spring set point is set to be equal to 1, which is equal to the stance leg spring set point.