Study of Swing model sensitivities to IC & params
Looking at the return map of the current system is not informative because the simulation stops at a certain time, rather than
when the foot hits the ground.  The return map is calculated from the error to the expected stance leg angle & length at the start of heel strike, and the velocity at the end of the simulation compared to the nominal gait.

Instead, propose this return map:

Simulate the step until the swing leg hits the ground.  Calc return map between stance leg initial states and stance leg state of next step.

This may be more informative about the stability of how the swing leg effects the trajectory of the pelvis mass, which is ultimately the real focus.