getSLIPstates now a method of class Swing.
The method does the same thing as the function, but also will simulate the free fall of the pelvis if the time of Swing simulation exceeds that of the SLIP data.  This happens if the Swing property useHSevent = 1, which will detect when the swingfoot hits the ground instead of stopping the simulation at the time when the next step "should" start according to a steady limit cycle.

The useHSevent is necessary to study perturbations and sensitivities of the model to parameters and ICs.