Get Eigvals of Return map as fcn of pert size
Extreme nonlinearities in the return map, so calculated a return map with differing magnitudes and signs to investigate how they are effected.
The result is that small perturbations in certain direction cause large changes in the state of the next step due to the foot barely missing the ground and then oscillating back upwards.
Possible ways to deal with this:
1)  Have a locking mechanism when the swing leg reaches full extension
2)  Include some kind of control mechanism such as inputting or removing energy