Added lower bound to impulsecoeff
In class Swing, the property impulsecoeff determines how much energy is transferred to the foot at take off.  This is supposed to model a yank on a string upwards to get the foot forwards.  During optimization, the coefficient can be set to negative which sends the foot shooting downwards.  I added a lower bound to optimizations using this parameter in method findLimitCycle of class Runner in order to prevent this.

Within class Runner are general rules for constraints on parameters to make them realistic: ie, springs, dampers, masses > 0.  There are also rules for other specific parameters of models.

In the future, it may be good to remove these class specific-constraint rules and add them directly to the classes, and have findLimitCycle check for these rules before starting the optimization