Created a swing model with spring b/w pelvis & foot
This is kind of a biarticular spring model.  The idea is that the set point of the pulley spring is forwards, so that it is biased towards moving the foot forwards.  Whereas with a regular knee spring, since the knee is supposed to be straight at both take off and landing,  there is nothign biasing it towards a landing configuration.

This may be important for stability.  In the knee model, if there is a perturbation which causes the foot barely to not hit the ground for the next step, the knee spring pulls up the foot for another cycle and causes a big nonlinearity in the return map.  While locking the knee is an alternative to having a biased spring, this model may give additional insight into the stability dynamics without adding the nonlinearity of locking the knee.