1D Param studies on Swing Set Point
For model swing, did several 1D parameter studies on the swing spring set point, allowing the optimizer to find a limit cycle by varying different sets of parameters.  Observations are written at the title of each cell.

Also reogranized structure of SwingStudy.m to automatically detect if the study has been done before and to load the results instead of computing the results each time unless the flag forcerun is set to 1.