Parameter Studies.
Did 1D and 2D parameter studies on the Swing model from a nominal gait with no impulse.  Studied khip, kswing, swingl, hipl, and impulsecoeff.

With some parameters, the foot goes through the floor.  With other set of parameters, the foot is retracted so close to the pelvis that it gives the hip DOF a nearly infinite inertia which makes it hard for the integrator and optimizer to find limit cycles.  The limit cycles tend to break down when the swing spring set point (swingl) is too small.