added property slip; use property slip for more efficient findLimitCycle
property slip contains a slip model whose gait char (speed, air frac, step length) match up with the properties speed, steplength, and airfrac of the SwingSLIP instance.

In findLimitCycle, it first checks to see that the SLIP model stored in property slip has these desired properties.  If it does not, it uses the slip model as the initial condition for an optimization to find a slip model that does.

After this, another optimization that solves for a limit cycle of the swing leg is done, assuming that the result from the SLIP model is accurate.